+++ src/seq/Seq.jl
@@ -47,6 +47,8 @@ export
     RNA_N,
     RNA_Gap,
     ACGU,
+    ACGUN,
+    isGC,
     iscompatible,
     isambiguous,
     iscertain,
@@ -174,29 +176,10 @@ import Combinatorics
 import IndexableBitVectors
 import Iterators
 importall Bio
-
-"""
-    alphabet(typ)
-
-Return an iterator of symbols of `typ`.
-
-`typ` is one of `DNA`, `RNA`, or `AminoAcid`.
-"""
-function alphabet end
-
-"""
-    gap(typ)
-
-Return the gap symbol of `typ`.
-
-`typ` is one of `DNA`, `RNA`, `AminoAcid`, or `Char`.
-"""
-function gap end
+importall BioSymbols
 
 gap(::Type{Char}) = '-'
 
-include("nucleicacid.jl")
-include("aminoacid.jl")
 include("alphabet.jl")
 include("bitindex.jl")
 include("sequence.jl")